backend/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
gunicorn==23.0.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
slowapi==0.1.9
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
openai==2.11.0
requests==2.32.5
